Changed some test methods to the proper ones, restructured the test folders and tested that the tests are successful
