+++ src/test/java/com/fti/softi/config/SecurityConfigTests.java
@@ -1,9 +1,10 @@
-package com.fti.softi;
+package com.fti.softi.config;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.security.test.context.support.WithMockUser;
 //import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -14,20 +15,20 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 public class SecurityConfigTests {
 
-    @Autowired
-    private MockMvc mockMvc;
+  @Autowired
+  private MockMvc mockMvc;
 
-    @Test
-    //@WithMockUser(username = "admin", roles = {"ADMIN"})
-    void testAdminAccess() throws Exception {
-        mockMvc.perform(get("/admin"))
-                .andExpect(status().isOk());
-    }
+  @Test
+  @WithMockUser(username = "admin", roles = {"ADMIN", "USER"})
+  void testAdminAccess() throws Exception {
+    mockMvc.perform(get("/admin"))
+      .andExpect(status().isOk());
+  }
 
-    @Test
-    //@WithMockUser(username = "user", roles = {"USER"})
-    void testUnauthorizedAccess() throws Exception {
-        mockMvc.perform(get("/admin"))
-                .andExpect(status().isForbidden());
-    }
+  @Test
+  @WithMockUser(username = "user", roles = {"USER"})
+  void testUnauthorizedAccess() throws Exception {
+    mockMvc.perform(get("/admin"))
+      .andExpect(status().isForbidden());
+  }
 }
