+++ src/test/resources/application-test.properties
@@ -1,0 +1,16 @@
+spring.datasource.url=jdbc:h2:mem:testdb;MODE=SQLite
+spring.datasource.driver-class-name=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.h2.console.enabled=true
+spring.jpa.hibernate.ddl-auto=create-drop
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+logging.level.org.springframework=DEBUG
+logging.level.com.yourpackage=DEBUG
+
+server.port=0
+
+spring.main.allow-bean-definition-overriding=true
