+++ src/test/resources/schema.sql
@@ -1,0 +1,5 @@
+CREATE TABLE "user" (
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    name VARCHAR(255),
+    email VARCHAR(255)
+);
