+++ src/test/resources/test.sql
@@ -1,0 +1,1 @@
+INSERT INTO "user" (name, email) VALUES ('John Doe', 'john.doe@example.com');
